LPS-66931 Move User groups admin data handlers to the correct packages
