+++ modules/apps/foundation/user-groups-admin/user-groups-admin-impl/src/main/java/com/liferay/user/groups/admin/internal/exportimport/data/handler/UserGroupsAdminPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.user.groups.admin.internal.lar;
+package com.liferay.user.groups.admin.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.DataLevel;
